fixed internationalization issues.
