+++ src/main/kotlin/net/masterthought/dlanguage/DLanguage.kt
@@ -1,13 +1,11 @@
 package net.masterthought.dlanguage
 
-import com.intellij.CommonBundle
 import com.intellij.lang.ASTNode
 import com.intellij.lang.Language
 import com.intellij.lang.ParserDefinition
 import com.intellij.lang.PsiParser
 import com.intellij.lexer.FlexAdapter
 import com.intellij.lexer.Lexer
-import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import com.intellij.psi.FileViewProvider
 import com.intellij.psi.PsiElement
@@ -19,50 +17,11 @@ import net.masterthought.dlanguage.parser.ParserWrapper
 import net.masterthought.dlanguage.psi.DLanguageFile
 import net.masterthought.dlanguage.psi.DLanguageTypes
 import net.masterthought.dlanguage.stubs.types.DFileStubElementType
-import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.NotNull
-import org.jetbrains.annotations.PropertyKey
-import java.lang.ref.Reference
-import java.lang.ref.SoftReference
-import java.util.*
 
 object DLanguage : Language("D")
 
 class DLanguageLexerAdapter : FlexAdapter(DLanguageLexer())
-
-object DLanguageBundle {
-
-    val log: Logger = Logger.getInstance(DLanguageBundle.javaClass)
-
-    var dLangBundle: Reference<ResourceBundle>? = null
-
-    @NonNls private const val BUNDLE_ID: String = "i18n"
-
-    init {
-        val locale = Locale.getDefault()
-        log.info("initialising D Language Bundle for locale: ${locale.toLanguageTag()}")
-    }
-
-    fun message(@PropertyKey(resourceBundle = BUNDLE_ID) key: String, vararg params: Any ): String {
-        log.debug("Getting message: {}, {}", key, params)
-
-        return CommonBundle.message(getBundle(), key, params)
-    }
-
-    private fun getBundle(): ResourceBundle {
-        var bundle: ResourceBundle? = null
-
-        if (dLangBundle != null) {
-            bundle = dLangBundle?.get()
-        }
-
-        if (bundle == null) {
-            bundle = ResourceBundle.getBundle(BUNDLE_ID)
-            dLangBundle = SoftReference<ResourceBundle>(bundle)
-        }
-        return bundle!!
-    }
-}
 
 class DLangParserDefinition : ParserDefinition {
     val WHITE_SPACES: TokenSet = TokenSet.create(TokenType.WHITE_SPACE)
